+++ src/main/java/ru/geek/mz/site/at/pages/Page.java
@@ -1,13 +1,13 @@
 package ru.geek.mz.site.at.pages;
 
-import com.google.gson.internal.$Gson$Preconditions;
-import org.openqa.selenium.By;
 import org.openqa.selenium.NotFoundException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
 import static java.lang.Integer.parseInt;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class Page extends BasePage {
@@ -32,16 +32,12 @@ public class Page extends BasePage {
 
     public Page checkHeader(String expected){
         assertEquals(expected, pageHeader.getText());
-//        assertEquals(expected, driver.findElement(headerSelector).getText());
         return this;
     }
 
-    public int checkCount(WebElement checkedElement){
+    public int checkedCount(WebElement checkedElement){
         return parseInt(checkedElement.getText().split("・")[1]);
     }
-
-
-
 
     public NavigationTab getNavigationTab(){
         return navigationTab;
@@ -68,30 +64,65 @@ public class Page extends BasePage {
     @FindBy(css = "ul.search-page-tabs > li > a[data-tab='companies']")
     private WebElement companiesTab;
 
-    public Page checkTabs(String tabTitle) {
+
+    public Page checkTab(String tabTitle) {
         switch (tabTitle) {
             case "professionTab": {
-                professionTab.isDisplayed();
+                assertThat((checkedCount(professionTab)), allOf(
+                        notNullValue(),
+                        greaterThan(2)
+                        )
+                );
+                System.out.println("profess - ok");
                 break;
             }
             case "coursesTab": {
-                coursesTab.isDisplayed();
+                assertThat(checkedCount(coursesTab), allOf(
+                        notNullValue(),
+                        greaterThan(15)
+                ));
+                System.out.println("courses ok");
                 break;
             }
             case "webinarsTab": {
-                webinarsTab.isDisplayed();
+                assertThat(checkedCount(webinarsTab), allOf(
+                        notNullValue(),
+                        greaterThan(180),
+                        lessThan(300)
+                ));
+                System.out.println("webinars ok");
                 break;
             }
-            case "blog": {
-                blogTab.isDisplayed();
+            case "blogTab": {
+                assertThat(checkedCount(blogTab), allOf(
+                        notNullValue(),
+                        greaterThan(300)
+                ));
+                System.out.println("blogs ok");
                 break;
             }
-            case "forum": {
-                forumTab.isDisplayed();
+            case "forumTab": {
+                assertThat(checkedCount(forumTab), allOf(
+                        notNullValue(),
+                        not(350)
+                ));
+                System.out.println("forum ok");
                 break;
             }
-            case "test": {
-                testsTab.isDisplayed();
+            case "testsTab": {
+                assertThat(checkedCount(testsTab), allOf(
+                        notNullValue(),
+                        not(0)
+                ));
+                System.out.println("test ok");
+                break;
+            }
+            case "companiesTab": {
+                assertThat(checkedCount(companiesTab), allOf(
+                        notNullValue(),
+                        not(0)
+                ));
+                System.out.println("companies ok");
                 break;
             }
             default: {
